Put intermediate directory for providing .ssh conf
.ssh directory must now be mapped onto /.mapped-ssh in the container for
the github push to work. Without that fix, the .ssh directory would be
indeed accessible, but ssh would refuse to work with the usual error
'Bad owner or permissions on /root/.ssh/config' or something like that.
Indeed, when mapping the current user into the container, then the files
would be owned by some unknown uid (from inside the container) instead
of the root user. And ssh refuses to use keys in this situation.